Add "runtime.inf" from David Given's original patch.
This file is a slightly altered version of the "runtime.s"
defined in David Given's original "Silicon Castles" patch,
as seen here:
http://www.ifarchive.org/if-archive/infocom/compilers/vbcc/vbcc_0.7-3.diff

This file provides the definition for functions such as `__xor`,
calls to which are generated by the vbccz compiler.

The difference between Given's "runtime.s" and this file is
that I have removed `cinvoke` (because we don't need to call
C code from Inform) and `__c_stack` (because it is already
defined in "Main.inf").