[z] Fix miscompilation of incompletely initialized and/or 1-sized char arrays.
Before this patch, the gen_ds() function would do the wrong thing with
arrays that are partially but not completely initialized. Consider:

    char x[2][4] = { "h", "w" };

The generated .s file looks like this:

    Array _x ->
     (104)
     (0)
     2
     (119)
     (0)
     2
    ;

After this patch, those two "2"s are properly replaced with "(0) (0)"s.

Furthermore, consider the role of reflower() in handling globals and statics
of type char.

    char c = 42;

    Array _c ->
      (42)
      (0)
    ;

Notice the extra "(0)" appended by reflower(). We need that extra "(0)"
because `Array _c -> (42);` would indicate an array of forty-two 0s,
instead of an array of one 42. But the same codepath miscompiles

    char c;

    Array _c ->
      1
      (0)
    ;

where `Array _c -> 1;` is actually what we meant. After this patch, the
codegen for `char c;` is `Array _c -> (0) (0);`, which works as intended.