Fix a major backend bug with "stackparamadjust".
To compile a call to a variadic function, the middle-end generates
a series of PUSHes followed by a CALL. It doesn't generate any POPs;
it leaves that up to the backend.

We do need to generate POPs to get the stack pointer (that's "xp",
not "sp") back to its old value. If we left "xp" unbalanced, we'd
have problems with diamond-shaped control flow:

    if (foo)
        printf("%d", 42);  /* decrements xp by 2 */
    localvar = 1;  /* we need to know localvar's offset relative
                      to xp, but now that depends on foo! */

Before this revision, David Given's algorithm was to assume that
every CALL should reset xp's offset to zero. But this isn't true!
Consider:

    printf("You scored %d out of a possible %d.\n", score(), 350);

This corresponds to an IC of

    PUSH 350
    CALL _score -> r0
    PUSH r0
    ASSIGN "%d %d" -> r0
    CALL _printf

Given's algorithm caused printf to "forget" any parameters that
followed a parameter containing a function call!