[z] Remove the hack for Inform bug C62105.
Actually it's still there, but under a switch (-patch-not) which is off
by default. This means we can compile C expressions containing "~" without
needing to provide a library function __not().

This code was working around a bug in Inform <6.30:
http://www.inform-fiction.org/patches/C62105.html
The bug is fixed in Inform 6.31 and 6.32, which are the versions of
Inform 6 that are currently available for Ubuntu Linux.