Fix a Clang diagnostic caused by a typo-bug in "declaration.c".
I suspect that this patch has no effect, because it seems like
any time `p->vtyp` is not FUNKT, `p->fi` ought to be NULL anyway,
so the first clause of this disjunction is never significant.
But Clang complains, and the fix is obvious, so let's apply it.

```
declaration.c:3094:16: warning: logical not is only applied to the left hand side of this comparison
      [-Wlogical-not-parentheses]
            if(!(p->vtyp->flags&NQ)==FUNKT||!p->fi||!p->fi->inline_asm)
               ^                   ~~
declaration.c:3094:36: warning: comparison of constant 15 with boolean expression is always false
      [-Wtautological-constant-out-of-range-compare]
            if(!(p->vtyp->flags&NQ)==FUNKT||!p->fi||!p->fi->inline_asm)
               ~~~~~~~~~~~~~~~~~~~~^ ~~~~~
```