Add inlining of puts/printf, and an "-opt-puts=" option.
I'm inserting this optimization right into the indep "ic.c", just because
there doesn't seem to be any elegant place to put it. (I don't think
HAS_LIBCALLS is quite what I'm looking for, unfortunately.)

The command-line option makes it possible to get the effect of
    #define puts print_indented
in Z-code. Of course we could already write exactly that in C, and the
compiler would handle it; but because the function call wasn't spelled
"puts", it wouldn't get the compressed-string-literal optimization.

You can supply a Z-machine function name to the -opt-puts= option; for
example "-opt-puts=print_indented". This will cause the C expression
    puts("hello")
to be compiled into
    print_indented("hello^")
on the Z-machine. The default behavior is equivalent to
"-opt-puts=print (string) ".

The single function call
    printf("one %s two %d three\n", arg1, arg2);
will be compiled into the equivalent of
    print "one ";
    printf("%s", arg1);
    print " two ";
    @print_num arg2;
    print " three^";

Printing integers via printf("%d") is still compiled into "print (int)"
regardless of the -opt-puts= option; and embedded Unicode characters will
still be printed via @print_unicode regardless of the -opt-puts= option.
This is fine for my current needs.