[z] Fix miscompilation of ~ << >> for LONG values.
Unary ~ seems like a straightforward oversight on Given's part; when
long arithmetic was added, the path for the KOMPLEMENT operator
was never implemented.

Notice a quirk of vbcc here: the IC for a left or right shift always forces
the right-hand operand to type INT! This means we actually have to check
for that case separately from all the other arithmetic operators, and that
that `__long_{asr,lsl,lsr}` routines have different signatures from all
the other `__long_*` routines in the runtime.