Fix a typo-bug in push_value() causing non-deterministic miscompilation.
We meant to check obj->val.vlong==0 "is this constant-expression
of the form &var as opposed to &var+42", but instead we were
checking obj->v->offset==0 "does var have a stack offset of zero",
which of course is nonsense for any EXTERN or STATIC variable.
Not only is it nonsense, but the "offset" field actually tends to
contain garbage, which means that the accompanying test case fails
nondeterministically.

The main effect of this patch is to replace many snippets of the form
    @add STATIC_main_227 (0) -> sp;
    @pull r0;
with snippets of the form
    @add STATIC_main_227 (0) -> r0;
but occasionally it goes the other way, and replaces
    @add STATIC_main_227 (0) -> r0;
with
    @add STATIC_main_227 (2) -> sp;
    @pull r0;
(or whatever the appropriate "vlong" field should be). These are the cases
that fix real miscompilations.