Fix an ICE and implement "char -> long" conversions.
Type conversions from CHAR to LONG were implemented in the compiler but
not in "runtime.inf"; the `__long_fromchar` routine was missing.

Type conversions from CHAR to UNSIGNED|LONG were not implemented in
the compiler, due to an oversight in a switch statement; this resulted
in an ICE.

Both of these bugs are now fixed.