Fix an inefficient-codegen bug in the indep find_local_reg().
I observed that given the IC

    GETRETURN -> v0
    SUB v0, 4 -> v1

where v0 and v1 are both unnamed (temporary) variables and v1 has already
been assigned to register 2 (r0 on the Z-machine backend), find_local_reg(v0)
decides that because the live-range of v0 overlaps with the SUB's assignment
to v1, therefore `savings[2]=INT_MIN` (i.e., let's try not to allocate v0
into r0). However, this is incorrect: the two live-ranges do not overlap,
because the SUB happens to coincide with the *end* of v0's live-range (which
find_local_reg() calls `fp`). If the "overlapping" assignment into a register
happens when `p==fp`, then that's fine and we should not set
`savings[abs(p->z.v->reg)]=INT_MIN` after all.

I believe the same logic is valid for the register-pair code guarded by
this same `if`. (The Z-machine backend doesn't use register pairs.)

This patch drastically improves codegen for

    int main() { return f(~g(~h())); }

from the original

        @call_vs2 _h xp r0 r1 r2 r3 r4 r5 -> r0;
        r1 = r0;
        @not r1 -> r0;
        @call_vs2 _g xp r0 r1 r2 r3 r4 r5 -> r0;
        r1 = r0;
        @not r1 -> r0;
        @call_vs2 _f xp r0 r1 r2 r3 r4 r5 -> r0;
        @ret r0;

to simply

        @call_vs2 _h xp r0 r1 r2 r3 r4 r5 -> r0;
        @not r0 -> r0;
        @call_vs2 _g xp r0 r1 r2 r3 r4 r5 -> r0;
        @not r0 -> r0;
        @call_vs2 _f xp r0 r1 r2 r3 r4 r5 -> r0;
        @ret r0;

because the register allocator no longer shies away from allocating all the
temporaries into the same register r0.