Fix wrong codegen in dangerous_IC() and a segfault in dump_obj().
The sense of dangerous_IC()'s return value had been flipped for some reason:
the compiler was actually trying to hoist divisions and dereferences whenever
possible, and *failing* to hoist all other operations. This leads to inefficient
codegen in general, but also to incorrectly hoisted divide-by-zeroes.
Even an expression as simple as `(d ? 1/d : 0)` would do the wrong thing,
if placed inside a loop at `-O2`.

The test case in this patch exposes the bug at `-O2` but not at `-O3`.
At `-O2`, it consistently generates a hoisted division by zero, which
corresponds to "fatal error: divide by zero" in the Bocfel interpreter.

Furthermore, when given the IC for `*(int*)0`, dump_obj() got confused
by the IC's `flags==(KONST|DREFOBJ)`. The old dump_obj() assumed that the
only time `(flags&KONST)` is when `(flags==KONST)`, so it would fall into
the codepath that assumed `(flags&VAR)`, dereference `obj->var` which would
be garbage, and segfault. This is now fixed.