Fix wrong codegen for variadic functions taking non-register parameters.
The vbccz calling convention passes up to 6 register-sized arguments
in r0 through r5, and then starts pushing things on the stack (xp).
This means that in a variadic function, there may be arguments on the
stack below (pushed-earlier-than) the "varargs" portion of the stack.
For some reason, the old code wasn't counting those arguments' sizes
toward the "varargs" offset. This is now fixed.