Fix a bug where r0 was not being marked as trashed by function calls.
The accompanying test case reproduces the bug, but only at -O0.

The compiler must be taught that `r0` is not a safe place to stash
temporary results across a function call, even though all other
registers are safe (because they aren't really registers; they're
local Inform variables). `r0` is special because the codegen for
a CALL actually assigns back into the local `r0`.

Ideally we might want to assign into the local `r0` only if the CALL
being codegenned is followed by a GETRETURN; and then we wouldn't
need to mark `r0` trashed in those other cases (e.g. functions returning
`void` or `long` or `struct`). But that's departing quite a bit from
the intuition of "registers" and "calling conventions" on which vbcc
is based, and would probably lead to even subtler bugs later on.