[vbcc] Fix an ICE due to incorrect handling of BITFIELD in type_expression2().
The attached test case produces an ICE even at -O0:

    >    struct S s; (s.f = 42) + 0;
    error 86 in line 10 of "ice.c": lvalue required for assignment
    >    struct S s; (s.f = 42) + 0;
    error 158 in line 10 of "ice.c": internal error 0 in line 1197 of file type_expr.c !!

The reason is that type_expression2() is responsible for setting the `lvalue`
flag correctly on each AST node; but it also performs some transformations on
the tree, such as in this case changing a CONTENT node to a BITFIELD node.
So the first time we call type_expression2() on this tree, it changes CONTENT
to BITFIELD and sets `lvalue=1`, all correctly. But the second time we call it
(triggered by alg_opt()'s removal of that redundant `+ 0`), since the CONTENT
node is no longer in the tree, type_expression2() preserves the BITFIELD node
but incorrectly fails to set `lvalue=1`.

The ICE is now fixed. However, I'm not sure if this patch might accidentally
set `lvalue=1` on nodes that shouldn't be lvalues.