[vbcc] Fix an ICE where `nocode` gets out of sync with reality.
A label or a switch-case can reset `nocode` from 1 to 0, which can cause
the `BRA` at the end of an `if` block to get generated... but by the time
if_statement() gets around to generating the corresponding `LABEL`,
it will have re-reset `nocode=cm`. This causes us to generate a branch
to an undefined label, anytime an `if` containing a label or switch-case
appears in a context that is `nocode` for some other reason. This bug
is now fixed.