[vbcc] Fix an ICE due to a bug in the CSE optimization for structs.
The provided test case ICEs in falign(), in add_tmp_var(), in cse(); because
the type of the temporary variable we're trying to create is a STRUCT type
that is not `typ->exact`; and it's not `typ->exact` because it was just
created from scratch by cse() calling new_typ().

The reason cse() creates this temporary variable is because cse() thinks
it's appropriate to eliminate a common subexpression (whose type happens to
be a struct type); and that's because num_exp() told it to. num_exp() thought
it was okay because `p->code==ASSIGN && (p->q1.flags & DREFOBJ)`; but that's
blatantly not sufficient, given that `q1typ(p)==STRUCT`.

My patch here simply makes num_exp() more conservative in this one case.
I don't claim that this patch is either necessary to fix this ICE, or
sufficient to fix all the bugs in this vicinity. I suspect there are more
add_tmp_var()-related bugs yet to be found.