Add an test case that ICEs in fix_local_allocreg().
The error message is:
```
error 158 in line -2004024531 of "ice6.c": internal error 0 in line 1139 of file regs.c !!
```

It clearly has something to do with that unnamed function parameter;
giving it a name fixes the ICE.

Part of the problem seems to be in cross_module_inline(); setting -O=2147467263
(that's "every bit but 16384") prevents the ICE, although then the compiler
produces an error message and still fails to compile the program.

```
$ VBCC=../vbcc ../vbcc/bin/vc +ppc -c99 -O=4098 -c ice6.c -o main.s
>    g(0);
error 39 in line 11 of "ice6.c": invalid types for assignment
1 error found!
```

TODO: investigate and fix.