[vbcc] Add an test case that ICEs on line 1216 of fix_local_allocreg()
It clearly has something to do with that unnamed function parameter;
giving it a name fixes the ICE.

Part of the problem seems to be in cross_module_inline(); setting -O=2147467263
(that's "every bit but 16384") prevents the ICE, although then the compiler
produces an error message and still fails to compile the program.

```
$ VBCC=../vbcc ../vbcc/bin/vc +ppc -c99 -O=4098 -c todo4.c
>    g(0);
error 39 in line 8 of "todo4.c": invalid types for assignment
1 error found!
```

TODO: investigate and fix.