[vbcc] Fix an ICE due to an unhandled codepath in inline_memcpy().
Code such as `memcpy(1,2,n)` causes an ICE in inline_memcpy()
because it doesn't know how to handle non-pointer arguments. It
shouldn't really need to handle them, but it also shouldn't ICE
just because the call is impossible to inline. I've changed the
return type of inline_memcpy() from void to int so that it can
report failure back to its caller.