[z] Pointer comparison on the Z-machine should be unsigned.
If pointer comparison is signed, then an array allocated at address 0x7FF0
will have `&arr[0] > &arr[0x20]`, which is incorrect. The unsigned comparison
is slightly more expensive, but pointer comparisons shouldn't happen very
much in real code anyway.

Also, this brings pointer comparisons into line with pointer
conversions-to-unsigned-long, which was already being done as if the
values were unsigned. (That is, we'd do zero-extension instead of
sign-extension.)