[vbcc] Fix a bug in get_clist_byte/get_scalar_byte
The symptom is a crash inside insert_const_memcpy.
I'm not 100% sure what this optimization is for; I think
it might be trying to propagate a constant initializer
from a variable to its first use. The problem is when
the first use is using a field-of-a-union which is bigger
than the *first* member of the union; in that case we
ask for the Nth byte of an initial value with only M bytes.
On a non-little-endian system that makes `j` negative,
and we loop forever. Solution: don't do that.